refactor(ai): consolidate provider-name checks in client.rs
Replace raw "openrouter" string literals with PROVIDER_OPENROUTER
constant, extract duplicated free-tier validation into shared
validate_openrouter_free_tier helper, and relocate is_free_model()
and resolve_anthropic_credential() from ai/mod.rs into client.rs
with compatibility re-exports preserving the external API.

Closes #1482
